Migrate cluster-overview and dependencies
