--- web/html/src/components/panels/PanelRow.js
@@ -1,20 +1,0 @@
-// @flow
-
-import * as React from 'react';
-
-type Props = {
-  className?: string,
-  children: React.Node,
-}
-
-function PanelRow(props: Props) {
-  return (
-    <div className={`row ${props.className != null ? props.className : ""}`}>
-      {props.children}
-    </div>
-  );
-}
-
-export {
-  PanelRow,
-};
