+++ web/html/src/components/panels/PanelRow.tsx
@@ -1,0 +1,12 @@
+import * as React from "react";
+
+type Props = {
+  className?: string;
+  children: React.ReactNode;
+};
+
+function PanelRow(props: Props) {
+  return <div className={`row ${props.className || ""}`}>{props.children}</div>;
+}
+
+export { PanelRow };
